requirements.txt
mlx-whisper
pyobjc-framework-Cocoa
pyobjc-framework-Quartz
sounddevice
soundfile
numpy
